Voxel shaders working

+++ assets/shaders/gBufferVoxel.vert
@@ -2,12 +2,12 @@
 layout (location = 0) in vec3 vertexPosition;
 layout (location = 1) in vec2 vertexTexCoord;
 layout (location = 2) in vec3 vertexNormal;
-layout (location = 3) in float vertexTextureId;
+layout (location = 3) in int vertexTextureId;
 
 out vec3 fragPosition;
 out vec2 fragTexCoord;
 out vec3 fragNormal;
-out float fragTextureId;
+out int fragTextureId;
 
 uniform mat4 matModel;
 uniform mat4 matView;
